ML_Model/requirements.txt
scikit-learn
pandas
numpy
streamlit
PyPDF2
langchain
langchain_community
openai
tiktoken
transformers
sentence_transformers
langchain_google_genai
streamlit
PyPDF2
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
chromadb